requirements.txt
requests
boto3
coloredlogs
pyyaml
GitPython